Add files via upload
This is the first draft of the Options Class, in which I create a simple class to instantiate options-like objects. These objects have built-in methods to calculate their greeks and their price. It also has an option strategy visualization tool, so the user can give a list of options as input and see the different payoff structures of the strategy for a customizable range of prices. 
Note that these options only account for one risk-free rate, so there's only one measure of rho. A second version might include more specific type of options that can specify their asset class.