Add files via upload
This new version of the Options Class includes a second interest rate for more generalized uses. Note that the Options.greeks() method now receives an argument to specify if it's supposed to return both of Rho's calculations (assuming that the option in question is of foreign currency). The graphic components are still the same and perfectly functional.