File description
This file contains a function that builds a binomial tree given a spot price, strike price, probabilities of upward and downward price movements, a risk-free rate, delta T, and whether the option to be priced is a call or a put. It outputs the correspondent option price and the whole structure of the tree, which is a series of accessible dictionaries that show both the price trajectories of the stock according to the given probabilities, and the value of the option at that node. 

Note that the function allows for any number of steps and provides access to any node of any step, accessing it like a regular dictionary.