Greek Visualizer
This code contains a function allowing users to see how traditional Greeks' values change with different spot prices of the underlying asset. It uses option-like objects created via the Options class. The function, on the other hand, requires only two arguments: the Greek to plot and the corresponding option.
Finally, this function is added to the finfun.py file.