Correction of minor bugs
This version corrects some minor bugs, like the function to print the Greeks (it was returning Gamma as Vega) and a minus sign damaging Rho's results. Moreover, now the code considers if the trader is going long or short to correctly change the Greeks' signs. It also gives an additional parameter to output the Greeks with a given number of decimals, for easier readability.